components/image-handler/requirements.txt
boto3==1.25.4
Pillow==9.3.0
